+++ service/src/main/java/uk/gov/hmcts/reform/fpl/service/email/content/CaseManagementOrderEmailContentProvider.java
@@ -16,8 +16,6 @@ import java.util.Map;
 import static uk.gov.hmcts.reform.fpl.utils.EmailNotificationHelper.buildSubjectLine;
 import static uk.gov.hmcts.reform.fpl.utils.EmailNotificationHelper.buildSubjectLineWithHearingBookingDateSuffix;
 import static uk.gov.hmcts.reform.fpl.utils.EmailNotificationHelper.formatCaseUrl;
-import static uk.gov.hmcts.reform.fpl.utils.PeopleInCaseHelper.getAllocatedJudgeName;
-import static uk.gov.hmcts.reform.fpl.utils.PeopleInCaseHelper.getAllocatedJudgeTitle;
 import static uk.gov.hmcts.reform.fpl.utils.PeopleInCaseHelper.getFirstRespondentLastName;
 import static uk.gov.service.notify.NotificationClient.prepareUpload;
 
@@ -68,8 +66,8 @@ public class CaseManagementOrderEmailContentProvider extends AbstractEmailConten
         return ImmutableMap.<String, Object>builder()
             .putAll(buildCommonCMONotificationParameters(caseDetails))
             .put("respondentLastName", getFirstRespondentLastName(caseData.getRespondents1()))
-            .put("judgeTitle", getAllocatedJudgeTitle(caseData.getAllocatedJudge()))
-            .put("judgeName", getAllocatedJudgeName(caseData.getAllocatedJudge()))
+            .put("judgeTitle", caseData.getAllocatedJudge().getAllocatedJudgeTitle())
+            .put("judgeName", caseData.getAllocatedJudge().getAllocatedJudgeName())
             .build();
     }
