+++ service/src/main/java/uk/gov/hmcts/reform/fpl/utils/PeopleInCaseHelper.java
@@ -1,14 +1,14 @@
 package uk.gov.hmcts.reform.fpl.utils;
 
-import uk.gov.hmcts.reform.fpl.model.*;
+import uk.gov.hmcts.reform.fpl.model.Representative;
+import uk.gov.hmcts.reform.fpl.model.Respondent;
+import uk.gov.hmcts.reform.fpl.model.RespondentParty;
 import uk.gov.hmcts.reform.fpl.model.common.Element;
 
 import java.util.List;
 import java.util.Objects;
 
 import static java.util.stream.Collectors.toList;
-import static uk.gov.hmcts.reform.fpl.enums.JudgeOrMagistrateTitle.MAGISTRATES;
-import static uk.gov.hmcts.reform.fpl.enums.JudgeOrMagistrateTitle.OTHER;
 
 public class PeopleInCaseHelper {
 
@@ -31,19 +31,5 @@ public class PeopleInCaseHelper {
             .map(representative -> String.format("%s\n%s", representative.getFullName(),
                 representative.getAddress().getAddressAsString(", ")))
             .collect(toList());
-    }
-
-    public static String getAllocatedJudgeTitle(Judge judge) {
-        if (judge.getJudgeTitle() == OTHER) {
-            return judge.getOtherTitle();
-        }
-        return judge.getJudgeTitle().getLabel();
-    }
-
-    public static String getAllocatedJudgeName(Judge judge) {
-        if (judge.getJudgeTitle() == MAGISTRATES) {
-            return judge.getJudgeFullName();
-        }
-        return judge.getJudgeLastName();
     }
 }
